Exports for stdcalls are usually decorated (with the notable exception of the windows api) like this: _someFuntionCall@4 where the numbers is the bytes of the required parameters.
This patch adds support for stdcall name decorations so that ffi users can specify just the funtion name (in this case someFunctionCall).  Its  also a helpful debugging tool - if you get the parameter types incorrectly ffi will throw an error about not being able to find the function name in the dll.